src/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
executing==2.0.1
flake8==7.0.0
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
Markdown==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
wcwidth==0.2.13
